Week-3-Learning Machine Learning Libraries
Download and Complete below exercises in Jupyter Notebook.

Each of them have set of exercises to complete. Please execute each of the exercise and check output generated and save the notebooks with all outputs genearated and then submit. 

Please do not submit the below starter notebooks as your homework without completing the exercises.